move tests into __tests__ for each folder
