+++ src/components/toggles/__tests__/FontWeight.test.js
@@ -1,4 +1,4 @@
-import FontWeight from './FontWeight';
+import FontWeight from '../FontWeight';
 import React from 'react';
 import Enzyme, { shallow, mount, render } from 'enzyme';
 import toJson from 'enzyme-to-json';
@@ -7,13 +7,13 @@ import 'jest-styled-components';
 
 Enzyme.configure({ adapter: new Adapter() });
 
-it('it correctly renders the FS toggle', () => {
+it('it correctly renders the FontWeight toggle', () => {
   const wrapper = shallow(<FontWeight />);
 
   expect(toJson(wrapper)).toMatchSnapshot();
 });
 
-it('has three toggles', () => {
+it('has 5 buttons', () => {
   const wrapper = mount(<FontWeight />);
   expect(wrapper.find('button')).toHaveLength(5);
 });
