+++ src/context/__tests__/BodyColorContext.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BodyColorContext, BodyColorProvider } from './BodyColorContext';
+import { BodyColorContext, BodyColorProvider } from '../BodyColorContext';
 import Enzyme, { mount } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 Enzyme.configure({ adapter: new Adapter() });
