+++ src/context/__tests__/ItalicContext.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ItalicContext, ItalicProvider } from './ItalicContext';
+import { ItalicContext, ItalicProvider } from '../ItalicContext';
 import Enzyme, { mount } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 Enzyme.configure({ adapter: new Adapter() });
