+++ src/context/__tests__/TextColorContext.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TextColorContext, TextColorProvider } from './TextColorContext';
+import { TextColorContext, TextColorProvider } from '../TextColorContext';
 import Enzyme, { mount } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 Enzyme.configure({ adapter: new Adapter() });
